Modularize the Talon interface further
