+++ CommonChickenRuntimeEngine/src/ccre/drivers/ctre/talon/TalonSRX.java
@@ -18,10 +18,6 @@
  */
 package ccre.drivers.ctre.talon;
 
-import ccre.channel.BooleanIO;
-import ccre.channel.FloatIO;
-import ccre.channel.FloatInput;
-import ccre.ctrl.ExtendedMotor;
 import ccre.ctrl.Faultable;
 
 /**
@@ -135,172 +131,21 @@ public interface TalonSRX {
      *
      * @return the ExtendedMotor for this Talon.
      */
-    public ExtendedMotor modMotor();
-
-    // other stuff
+    public TalonExtendedMotor modMotor();
 
     /**
-     * Provides the enablement state of this ExtendedMotor.
+     * Accesses a representation of the feedback data of the Talon.
      *
-     * @return a BooleanIO representing the enable state of the Talon.
+     * @return the Feedback module.
      */
-    public BooleanIO asEnable();
+    public TalonFeedback modFeedback();
 
     /**
-     * Provides the bus voltage of the Talon, in volts.
+     * Accesses a representation of the general configuration of the Talon.
      *
-     * The bus voltage is the input voltage of the Talon, generally the same as
-     * the battery voltage.
-     *
-     * @return a FloatInput representing the bus voltage of the Talon.
+     * @return the General Configuration module.
      */
-    public FloatInput getBusVoltage();
-
-    /**
-     * Provides the output voltage of the Talon, in volts.
-     *
-     * The output voltage is the actual voltage across the terminals of the
-     * driven motor.
-     *
-     * @return a FloatInput representing the output voltage of the Talon.
-     */
-    public FloatInput getOutputVoltage();
-
-    /**
-     * Provides the output current of the Talon, in amps.
-     *
-     * The output current is the actual current passing through the Talon.
-     *
-     * @return a FloatInput representing the current going through the Talon.
-     */
-    public FloatInput getOutputCurrent();
-
-    /**
-     * Provides the current position of the Talon's active sensor, in the
-     * engineering units for the specific sensor.
-     *
-     * @return a FloatIO representing the sensor position of the currently
-     * active sensor.
-     */
-    public FloatIO getSensorPosition();
-
-    /**
-     * Provides the current velocity of the Talon's active sensor, in the
-     * engineering units for the specific sensor.
-     *
-     * @return a FloatInput representing the sensor velocity of the currently
-     * active sensor.
-     */
-    public FloatInput getSensorVelocity();
-
-    /**
-     * Provides the current throttle fraction of the Talon, from -1.0 to 1.0.
-     * This factor scales the active output voltage of the Talon.
-     *
-     * @return a FloatInput representing the Talon throttle.
-     */
-    public FloatInput getThrottle();
-
-    /**
-     * Provides the closed loop error reported by the Talon, in units based on
-     * the current sensor.
-     *
-     * @return a FloatInput representing the closed loop error.
-     */
-    public FloatInput getClosedLoopError();
-
-    // modes
-
-    /**
-     * Configures the Talon to follow the specified other Talon rather than
-     * follow a specific control mode of its own.
-     *
-     * The throttle of this Talon will replicate the other Talon's throttle as
-     * closely as possible.
-     *
-     * @param talonID the ID of the Talon to follow.
-     */
-    public void activateFollowerMode(int talonID);
-
-    /**
-     * Configures the Talon to follow the specified other Talon rather than
-     * follow a specific control mode of its own.
-     *
-     * The throttle of this Talon will replicate the other Talon's throttle as
-     * closely as possible.
-     *
-     * @param talon the Talon to follow.
-     */
-    public default void activateFollowerMode(TalonSRX talon) {
-        activateFollowerMode(talon.getDeviceID());
-    }
-
-    // public void activateMotionProfileMode(); TODO
-
-    // configuration
-
-    /**
-     * Configures whether or not the sensor reading and/or the output reading
-     * are negated.
-     *
-     * @param flipSensor whether or not the sensor's direction should be
-     * reversed.
-     * @param flipOutput whether or not the output's direction should be
-     * reversed.
-     */
-    public void configureReversed(boolean flipSensor, boolean flipOutput);
-
-    /**
-     * Configures the allowable closed loop error on the Talon.
-     *
-     * @param allowableCloseLoopError the maximum allowable error.
-     */
-    public void configureAllowableClosedLoopError(float allowableCloseLoopError);
-
-    /**
-     * Configures the rates at which the two common frames will be updated. Read
-     * the Talon SRX documentation for details.
-     *
-     * @param millisGeneral the millisecond period of the "general" frame.
-     * @param millisFeedback the millisecond period of the "feedback" frame.
-     */
-    public void configureGeneralFeedbackUpdateRate(int millisGeneral, int millisFeedback);
-
-    /**
-     * Configures the maximum output voltage of the Talon.
-     * <code>forwardVoltage</code> should be in the range 0 to 12, and
-     * <code>reverseVoltage</code> should be in the range 0 to -12.
-     *
-     * @param forwardVoltage the maximum forward voltage.
-     * @param reverseVoltage the maximum reverse voltage.
-     */
-    public void configureMaximumOutputVoltage(float forwardVoltage, float reverseVoltage);
-
-    /**
-     * Configures the nominal output voltage of the Talon.
-     * <code>forwardVoltage</code> should be in the range 0 to 12, and
-     * <code>reverseVoltage</code> should be in the range 0 to -12.
-     *
-     * @param forwardVoltage the nominal forward voltage.
-     * @param reverseVoltage the nominal reverse voltage.
-     */
-    public void configureNominalOutputVoltage(float forwardVoltage, float reverseVoltage);
-
-    // other
-
-    /**
-     * Provides the current temperature of the Talon, in degrees Celsius.
-     *
-     * @return a FloatInput representing the temperature of the Talon.
-     */
-    public FloatInput getTemperature();
-
-    /**
-     * Queries the firmware version of the Talon.
-     *
-     * @return the firmware version of the Talon, if it has been received.
-     */
-    public long GetFirmwareVersion();
+    public TalonGeneralConfig modGeneralConfig();
 
     /**
      * Fetches the device ID of the Talon. This was the ID used to create the
@@ -309,17 +154,4 @@ public interface TalonSRX {
      * @return the device ID, in the range 0 - 62.
      */
     public int getDeviceID();
-
-    /**
-     * Provides control over whether the Talon SRX is set to Brake or Coast. It
-     * will be true when the Talon is set to Brake and false when the Talon is
-     * set to Coast.
-     *
-     * Once this is changed, it will override anything configured on the roboRIO
-     * inspection website.
-     *
-     * @return a BooleanIO representing whether or not the Talon is set to Brake
-     * mode.
-     */
-    public BooleanIO getBrakeNotCoast();
 }
